partialSchedule: use templated isl types
An explicit cast is introduced in partialSchedule,
which will be removed after the schedule tree
has been converted to templated isl types.
Automatic type deduction is temporarily turned off in
promoteToSharedBelow and promoteToRegistersBelow to avoid having to convert
these entire functions together with partialSchedule.
It will be reintroduced when collectMappingsTo
is converted to templated isl types.