prefixSchedule: use templated isl types
An explicit cast is introduced in prefixSchedule,
which will be removed after the schedule tree
has been converted to templated isl types.
Automatic type deduction is temporarily turned off in
collectDomain and unroll to avoid having to convert
these entire functions together with prefixSchedule.
It will be reintroduced when the schedule tree
is converted to templated isl types.